chat_gpt/requirements.txt
fastapi==0.104.1
uvicorn==0.15.0
pytest==6.2.5
requests==2.26.0
pytest-asyncio==0.15.1
pytest-cov==3.0.0
flake8==4.0.1
openai==1.3.7
redis==3.5.3
python-dotenv